A bunch of logging and tracing.

+++ rust-zerotier-service/src/log.rs
@@ -20,34 +20,38 @@ use std::sync::Mutex;
 use chrono::Datelike;
 
 struct LogIntl {
+    prefix: String,
+    path: String,
     file: Option<File>,
     cur_size: u64,
     max_size: usize,
     log_to_stderr: bool,
 }
 
+/// It's big it's heavy it's wood.
 pub(crate) struct Log {
-    prefix: String,
-    path: String,
     inner: Mutex<LogIntl>,
 }
 
 impl Log {
     const MIN_MAX_SIZE: usize = 1024;
 
+    /// Construct a new logger.
+    /// If path is empty logs will not be written to files. If log_to_stderr is also
+    /// false then no logs will be output at all.
     pub fn new(path: &str, max_size: usize, log_to_stderr: bool, prefix: &str) -> Log {
         let mut p = String::from(prefix);
         if !p.is_empty() {
             p.push(' ');
         }
         Log{
-            prefix: p,
-            path: String::from(path),
             inner: Mutex::new(LogIntl {
+                prefix: p,
+                path: String::from(path),
                 file: None,
                 cur_size: 0,
                 max_size: if max_size < Log::MIN_MAX_SIZE { Log::MIN_MAX_SIZE } else { max_size },
-                log_to_stderr: log_to_stderr,
+                log_to_stderr,
             }),
         }
     }
@@ -61,59 +65,59 @@ impl Log {
     }
 
     pub fn log<S: AsRef<str>>(&self, s: S) {
-        // Output FATAL errors to stderr.
+        let mut l = self.inner.lock().unwrap();
+
         let ss: &str = s.as_ref();
         if ss.starts_with("FATAL") {
             eprintln!("{}", ss);
         }
+        let log_line = format!("{}[{}] {}\n", l.prefix.as_str(), chrono::Local::now().format("%Y-%m-%d %H:%M:%S").to_string(), ss);
 
-        let mut l = self.inner.lock().unwrap();
+        if l.path.len() > 0 {
+            if l.file.is_none() {
+                let mut f = OpenOptions::new().read(true).write(true).create(true).open(l.path.as_str());
+                if f.is_err() {
+                    return;
+                }
+                let mut f = f.unwrap();
+                let eof = f.seek(SeekFrom::End(0));
+                if eof.is_err() {
+                    return;
+                }
+                l.cur_size = eof.unwrap();
+                l.file = Some(f);
+            }
 
-        // If the file isn't open, open or create and seek to end.
-        if l.file.is_none() {
-            let mut f = OpenOptions::new().read(true).write(true).create(true).open(self.path.as_str());
-            if f.is_err() {
-                return;
+            if l.max_size > 0 && l.cur_size > l.max_size as u64 {
+                l.file = None;
+                l.cur_size = 0;
+
+                let mut old_path = l.path.clone();
+                old_path.push_str(".old");
+                let _ = std::fs::remove_file(old_path.as_str());
+                let _ = std::fs::rename(l.path.as_str(), old_path.as_str());
+                let _ = std::fs::remove_file(l.path.as_str()); // should fail
+
+                let mut f = OpenOptions::new().read(true).write(true).create(true).open(l.path.as_str());
+                if f.is_err() {
+                    return;
+                }
+                l.file = Some(f.unwrap());
             }
-            let mut f = f.unwrap();
-            let eof = f.seek(SeekFrom::End(0));
-            if eof.is_err() {
-                return;
+
+            let f = l.file.as_mut().unwrap();
+            let e = f.write_all(log_line.as_bytes());
+            if e.is_err() {
+                eprintln!("ERROR: I/O error writing to log: {}", e.err().unwrap().to_string());
+                l.file = None;
+            } else {
+                let _ = f.flush();
+                l.cur_size += log_line.len() as u64;
             }
-            l.cur_size = eof.unwrap();
-            l.file = Some(f);
         }
 
-        // If there is a maximum size limit configured, rotate if exceeded.
-        if l.max_size > 0 && l.cur_size > l.max_size as u64 {
-            l.file = None;
-            l.cur_size = 0;
-
-            let mut old_path = self.path.clone();
-            old_path.push_str(".old");
-            let _ = std::fs::remove_file(old_path.as_str());
-            let _ = std::fs::rename(self.path.as_str(), old_path.as_str());
-            let _ = std::fs::remove_file(self.path.as_str()); // should fail
-
-            let mut f = OpenOptions::new().read(true).write(true).create(true).open(self.path.as_str());
-            if f.is_err() {
-                return;
-            }
-            l.file = Some(f.unwrap());
-        }
-
-        let log_line = format!("{}[{}] {}\n", self.prefix.as_str(), chrono::Local::now().format("%Y-%m-%d %H:%M:%S").to_string(), ss);
         if l.log_to_stderr {
             stderr().write_all(log_line.as_bytes());
-        }
-        let f = l.file.as_mut().unwrap();
-        let e = f.write_all(log_line.as_bytes());
-        if e.is_err() {
-            eprintln!("ERROR: I/O error writing to log: {}", e.err().unwrap().to_string());
-            l.file = None;
-        } else {
-            let _ = f.flush();
-            l.cur_size += log_line.len() as u64;
         }
     }
 }
